requirements.txt
fastapi
psycopg2
asyncpg
sqlalchemy
databases
uvicorn
pytest
requests
openpyxl
python-jose
passlib
python-multipart
asyncio
bcrypt